requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
idna==3.4
iniconfig==1.1.1
kombu==5.2.4
packaging==22.0
pluggy==1.0.0
prompt-toolkit==3.0.36
pycodestyle==2.10.0
PyJWT==2.6.0
pytest==7.2.0
pytz==2022.7
redis==4.4.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
